ECOMM-156 refactoring: apdate api files structure
