+++ src/components/Router/Router.ts
@@ -1,8 +1,8 @@
 import type { Store } from '../Store/store';
 import { initLocation, changeLocation } from '../Store/slices/locationSlice';
-import { initAuth } from '../Store/slices/authSlice';
+import { initAuth, updateAuth } from '../Store/slices/authSlice';
 import type LocalStorage from '../LocalStorage/LocalStorage';
-import { getProductDetailsById } from '../Api/product';
+import { getProductDetailsById } from '../Api/rest/product';
 import { notifyError } from '../../utils/notify/notify';
 import { selectProduct } from '../Store/slices/productSlice';
 import { activeCartBindAction } from '../Store/store';
@@ -41,9 +41,7 @@ class Router {
     this.store = store;
 
     window.addEventListener('DOMContentLoaded', () => {
-      const localState = localStorage.loadState();
-      if (localState) this.store.dispatch(initAuth(localState.auth));
-      // this.store.dispatch(initAuth());
+      this.store.dispatch(initAuth());
       this.handleLocation('INIT_LOCATION');
     });
