+++ src/components/Store/slices/authSlice.ts
@@ -1,17 +1,9 @@
 import { createSlice, PayloadAction, createAsyncThunk } from '@reduxjs/toolkit';
 import type { Customer, CustomerDraft, CustomerUpdate, CustomerChangePassword } from '@commercetools/platform-sdk';
 import type { AuthState } from '../../../dto/types';
-import {
-  getToken,
-  getCustomerByToken,
-  createCustomer,
-  updateCustomerById,
-  updateCustomerPassword,
-  loginCustomer,
-} from '../../Api/rest/auth';
+import { createCustomer, updateCustomerById, updateCustomerPassword } from '../../Api/rest/auth';
 import { notifyError, notifyInfo } from '../../../utils/notify/notify';
 import { AuthPayload } from '../../../dto/types';
-import { meLogin } from '../../Api/rest/me';
 
 const signup = createAsyncThunk('auth/signup', async (payload: CustomerDraft) => {
   return createCustomer(payload)
@@ -28,22 +20,6 @@ const signup = createAsyncThunk('auth/signup', async (payload: CustomerDraft) =>
       notifyError(String(error.message)).showToast();
     });
 });
-
-// const login = createAsyncThunk('auth/login', async (payload: AuthPayload) => {
-//   return meLogin(payload)
-//     .then((response) => {
-//       if (response.statusCode !== 200) {
-//         let message = '';
-//         if ('message' in response) message = String(response.message);
-//         throw new Error(message);
-//       }
-//       notifyInfo('Successful login!').showToast();
-//       return response.body.customer;
-//     })
-//     .catch((error) => {
-//       notifyError(String(error.message)).showToast();
-//     });
-// });
 
 const updatePassword = createAsyncThunk('auth/updatePassword', async (payload: CustomerChangePassword) => {
   return updateCustomerPassword(payload)
@@ -90,11 +66,7 @@ const initialState: AuthState = {
   isEmailVerified: false,
   authenticationMode: '',
   inProgress: false,
-  tokenCache: {
-    token: null,
-    expirationTime: null,
-    refreshToken: null,
-  },
+  error: '',
 };
 
 const authSlice = createSlice({
@@ -104,8 +76,8 @@ const authSlice = createSlice({
     // initAuth(state) {
     //   Object.assign(state, {});
     // },
-    initAuth(state: AuthState, action: PayloadAction<AuthState>) {
-      Object.assign(state, { ...action.payload });
+    initAuth(state: AuthState) {
+      Object.assign(state, {});
     },
 
     updateAuth(state: AuthState, action: PayloadAction<Customer>) {
